requirements.txt
python-telegram-bot[job-queue]==20.8
SQLAlchemy>=2.0
psycopg[binary]
tzdata
